Quick navbar fix so that it is semi functional
